Initial implementation of Cro webapp forms
A mechanism for removing some of the tedium from working with HTML
forms in a web application. A form is declared as a class that does a
Cro::WebApp::Form role. The form can be rendered as HTML, and have its
data populated either like a normal object (`.new`) or by parsing it out
of a Cro body object for www-url-encoded data (later, multipart support
will also be added). The template language gains a prelude, which offers
a `form` function that renders a form into HTML.

Currently, it's very primitive, and also fairly incomplete; this is just
the initial building blocks. A layer will also later be built atop of
the forms to provide sugar for common CRUD scenarios.