Don't blow up in `.foo` on a missing hash key
If it's being used to index a hash, we assume that a soft failure is
desirable if the key does not exist. Having written quite a lot of
templates, this seems to be a better default. It will still blow up on a
non-associative (that is, an object-based model), however; the figuring
is that objects have a constant shape even if some methods return empty
results, whereas with hashes it's common to simply not even set the key
if the value would be undefined. Resolves #17 and fixes #15, which is
yet another example of a template I wrote that blew up for no good
reason due to the previous semantics.