Add :test parameter to template-location
This patch adds a `test` parameter to be passed to template-location
that works in a similar way to the parameter of the same name for
[dir].

By default, this will be set to ignore files and directories
whose basename has a leaing period (hidden files in UNIX systems).
This change in behaviour makes it so that, for example, working in
a cro project with an editor that ordinarily creates hidden swap
files (like vim) will not make this function automatically die with
a hard-to-understand error about malformed UTF-8 characters.

This is a change in behaviour, but it is unlikely that the old
behaviour was expected. Users that do want to match _all_ files
can, like with the [dir] routine, set this to `*`.

This parameter only takes effect if the `compile-all` flag is set.

[dir]: https://docs.raku.org/routine/dir