Implement templates from resources
This builds upon the recent router plugins and `resources-from` features
in Cro::HTTP to also provide a way to serve templates from a module's
resources. Furthermore, it makes `template-location` inside of a `route`
block only apply to the routes into that block and those included into
it. This is a potentially breaking change, however also an increase in
consistency. A `template-location` outside of any `route` block still
works, and will function as adding the location to a set of global
template locations.