Introduce template parts
The feature is primarily aimed at allowing for data for common elements
of pages to be produced and obtained without needing to explicitly pass
it in to every template call (think a header that has an indication of
if a user is currently logged in, or about their shopping cart content).
A `route` block integration in a subsequent commit will allow for that.
This one gets the basic mechanism in place, testing it by explicitly
providing arguments for the parts.

A further application of this is that the special part MAIN can be used
to obtain the initial argument (or argument capture) for rendering the
template. This will provide a way for folks to bind variables at the
"top level", in more complex cases where the topic approach doesn't
scale so well.