Implement route block template part providers
While one can explicitly provide template part data in the `template`
call, it will often be more convenient to avoid repetition and write a
`template-part` provider in the `route` block. This will be called when
the template part is encountered during rendering. It may optionally
have a parameter, which will be bound to `request.auth` so long as the
type matches; said type matching allows for different providers for
logged in or not, or different levels of user, or just to pull data out
of a session (for example, shopping cart content).