Add a way to parse a template from source
For those who might want to store the templates in some way other than
files or resources, for example in a database. We won't be providing
any template repositories for that in Cro::WebAPP, but this should be
enough for those who want them to implement such functionality as a
module.