Silence some warnings in escaping
If the object passed to the template has undefined keys (or methods that
return Nil,) then there were warnings produced from the escaping. For an
Associative this could potentially be worked around at source, but for
an object that has methods that may return Nil ( say an ORM row object,)
it's a bit less tractable.