Honour route block template locations in `use`
Previously we only paid attention to the `template-location` that is
declared as part of the `route` block (or templates from resources) in a
direct resolution of a template, but didn't pass these down in order for
them to be used when one template `use`d another. With this change, we
do so. Fixes #62.